Committing all fetch/filter/display files.
Not committing data files although they are
so format specific they are absolutely necessary.
It might be that I should check these in just as
a small subset - although probably not US which is
really another project anyway.

Be highly aware that I re-partitioned all of these
files and have not subsequently run them so this
will not currently work until completely mapped out
in terms of data dumps, intermediary files, and
locations.

Of course, there is also the question of automation
in the (essentially) ETL files.

The javascript is a monolith without structured form.